+++ channeldb/channel.go
@@ -1846,7 +1846,7 @@ func (c *OpenChannel) MarkShutdownSent(info *ShutdownInfo) error {
 // shutdownInfoKey.
 func (c *OpenChannel) storeShutdownInfo(info *ShutdownInfo) error {
 	var b bytes.Buffer
-	err := info.encode(&b)
+	err := encodeShutdownInfo(info, &b)
 	if err != nil {
 		return err
 	}
@@ -4817,40 +4817,17 @@ func DKeyLocator(r io.Reader, val interface{}, buf *[8]byte, l uint64) error {
 
 // ShutdownInfo contains various info about the shutdown initiation of a
 // channel.
-type ShutdownInfo struct {
-	// DeliveryScript is the address that we have included in any previous
-	// Shutdown message for a particular channel and so should include in
-	// any future re-sends of the Shutdown message.
-	DeliveryScript tlv.RecordT[tlv.TlvType0, lnwire.DeliveryAddress]
-
-	// LocalInitiator is true if we sent a Shutdown message before ever
-	// receiving a Shutdown message from the remote peer.
-	LocalInitiator tlv.RecordT[tlv.TlvType1, bool]
-}
+type ShutdownInfo = cstate.ShutdownInfo
 
 // NewShutdownInfo constructs a new ShutdownInfo object.
 func NewShutdownInfo(deliveryScript lnwire.DeliveryAddress,
 	locallyInitiated bool) *ShutdownInfo {
 
-	return &ShutdownInfo{
-		DeliveryScript: tlv.NewRecordT[tlv.TlvType0](deliveryScript),
-		LocalInitiator: tlv.NewPrimitiveRecord[tlv.TlvType1](
-			locallyInitiated,
-		),
-	}
+	return cstate.NewShutdownInfo(deliveryScript, locallyInitiated)
 }
 
-// Closer identifies the ChannelParty that initiated the coop-closure process.
-func (s ShutdownInfo) Closer() lntypes.ChannelParty {
-	if s.LocalInitiator.Val {
-		return lntypes.Local
-	}
-
-	return lntypes.Remote
-}
-
-// encode serialises the ShutdownInfo to the given io.Writer.
-func (s *ShutdownInfo) encode(w io.Writer) error {
+// encodeShutdownInfo serialises the ShutdownInfo to the given io.Writer.
+func encodeShutdownInfo(s *ShutdownInfo, w io.Writer) error {
 	records := []tlv.Record{
 		s.DeliveryScript.Record(),
 		s.LocalInitiator.Record(),
